+++ data/meta_filtered.json
@@ -1,0 +1,82 @@
+[
+  {
+    "title": "Watch This 1961 Chevrolet Impala Lowrider Dance | American tuned ft. Rob Dahm",
+    "channelId": "UCjOl2AUblVmg2rA_cRgZkFg",
+    "channelTitle": "Top Gear",
+    "thumbnailUrl": "https://i.ytimg.com/vi/rftjn89lAew/mqdefault.jpg",
+    "publishedAt": "2023-02-28T16:00:23Z",
+    "viewCount": "36316",
+    "likeCount": "1266",
+    "commentCount": "78"
+  },
+  {
+    "title": "2023 Corvette Z06 vs Ferrari 458 vs Porsche GT3 | Top Gear",
+    "channelId": "UCjOl2AUblVmg2rA_cRgZkFg",
+    "channelTitle": "Top Gear",
+    "thumbnailUrl": "https://i.ytimg.com/vi/bNkM-z3cPW8/mqdefault.jpg",
+    "publishedAt": "2023-02-26T16:00:16Z",
+    "viewCount": "482183",
+    "likeCount": "10972",
+    "commentCount": "1687"
+  },
+  {
+    "title": "Singer Factory Tour: How The Most Beautiful Porsches In The World Are Restored | Top Gear",
+    "channelId": "UCjOl2AUblVmg2rA_cRgZkFg",
+    "channelTitle": "Top Gear",
+    "thumbnailUrl": "https://i.ytimg.com/vi/2MSyHu9bMPo/mqdefault.jpg",
+    "publishedAt": "2023-02-23T16:45:08Z",
+    "viewCount": "556768",
+    "likeCount": "14093",
+    "commentCount": "589"
+  },
+  {
+    "title": "Honda Civic Type R Generations DRAG RACE",
+    "channelId": "UCUhFaUpnq31m6TNX2VKVSVA",
+    "channelTitle": "carwow",
+    "thumbnailUrl": "https://i.ytimg.com/vi/M46C6spmiJg/mqdefault.jpg",
+    "publishedAt": "2023-03-04T09:05:25Z",
+    "viewCount": "276638",
+    "likeCount": "18953",
+    "commentCount": "2214"
+  },
+  {
+    "title": "BMW X7 v Range Rover: ULTIMATE luxury test!",
+    "channelId": "UCUhFaUpnq31m6TNX2VKVSVA",
+    "channelTitle": "carwow",
+    "thumbnailUrl": "https://i.ytimg.com/vi/G2thNN9nVN8/mqdefault.jpg",
+    "publishedAt": "2023-03-03T12:10:46Z",
+    "viewCount": "537747",
+    "likeCount": "19921",
+    "commentCount": "1845"
+  },
+  {
+    "title": "Can a Porsche beat a Skoda RS OFF-ROAD?",
+    "channelId": "UCUhFaUpnq31m6TNX2VKVSVA",
+    "channelTitle": "carwow",
+    "thumbnailUrl": "https://i.ytimg.com/vi/_gTKcqp95Iw/mqdefault.jpg",
+    "publishedAt": "2023-03-01T11:04:13Z",
+    "viewCount": "507002",
+    "likeCount": "19167",
+    "commentCount": "1236"
+  },
+  {
+    "title": "New Toyota GR86 v Mk4 Supra: DRAG RACE",
+    "channelId": "UCUhFaUpnq31m6TNX2VKVSVA",
+    "channelTitle": "carwow",
+    "thumbnailUrl": "https://i.ytimg.com/vi/52lxhZgFrG8/mqdefault.jpg",
+    "publishedAt": "2023-02-28T11:07:29Z",
+    "viewCount": "788331",
+    "likeCount": "36932",
+    "commentCount": "3036"
+  },
+  {
+    "title": "BMW M3 Touring v Audi RS6: DRAG RACE",
+    "channelId": "UCUhFaUpnq31m6TNX2VKVSVA",
+    "channelTitle": "carwow",
+    "thumbnailUrl": "https://i.ytimg.com/vi/w0OPOSf1wlE/mqdefault.jpg",
+    "publishedAt": "2023-02-25T09:06:06Z",
+    "viewCount": "2102570",
+    "likeCount": "68711",
+    "commentCount": "7549"
+  }
+]
