+++ scraper/src/main/java/ch/unisg/senty/scraper/application/ScraperService.java
@@ -1,0 +1,18 @@
+package ch.unisg.senty.scraper.application;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+import org.springframework.stereotype.Component;
+
+import ch.unisg.senty.scraper.domain.Comment;
+
+@Component
+public class InventoryService {
+
+  /**
+   * Scrape through json mock data file and produce some juicy messages
+   */
+
+
+}
