--- scraper/src/main/java/ch/unisg/senty/scraper/domain/Item.java
@@ -1,25 +1,0 @@
-package ch.unisg.senty.scraper.domain;
-
-public class Item {
-
-  private String articleId;
-  private int amount;
-
-  public Item setArticleId(String articleId) {
-    this.articleId = articleId; 
-    return this;
-  }
-
-  public int getAmount() {
-    return amount;
-  }
-
-  public Item setAmount(int amount) {
-    this.amount = amount;
-    return this;
-  }
-
-  public String getArticleId() {
-    return articleId;
-  }
-}
